Add R packages and fix Python packages
The conda package for tensorflow was pulling in the GPU/CUDA version and
bloating the image. This commit specifies the cpu based tensorflow. It
also removes rise for a (hopefully) temporary bug where rise can't find
tornado.

Adding R packages in use on the callysto hub along with startup config
file and some other bits and pieces.

Signed-off-by: Ian Allison <iana@pims.math.ca>