Reduce impact of healthcheck and tweak lab notification config
The healthcheck is quite aggressive if you have a large number of users
so turn it down a bit. We'll monitor for problems and `HEALTHCHECK NONE`
if necessary (as we do on syzygy). The announcements/notification
toaster popup isn't quite right for our use case so we will disable
it for now.

We also updated the lock file so there are various package updates

Signed-off-by: Ian Allison <iana@pims.math.ca>